# backend/services/gateway/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: cportal-lrfe API (through the gateway)
  version: 0.1.0
  description: >-
    Every endpoint the browser can reach, at its /api/... path. The gateway only forwards requests: each service checks the token and
    permissions itself. Service-only endpoints are listed too, but answer 401 to a user token.


    In development, `ng serve` proxies /api to the gateway (angular-app/proxy.conf.json), so the paths work from the app's origin as they
    are.
servers:
  - url: http://localhost:3500
    description: gateway (local development)
...
paths:
...
  /api/catalogue:
    get:
      tags:
        - identity · Roles
      summary: Permission catalogue
      description: |-
        `{ groups, perms, roles }`: permission groups, every permission with its label, and the roles.

        **Access:** Any signed-in user.
      operationId: identity_getCatalogue
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
...
  /api/documents:
    post:
      tags:
        - edrms · Filing
      summary: File a document (intake only)
      description: >-
        multipart/form-data with a `meta` JSON field FIRST, then `file`. 201 `{ document, created: true }`, or 200 `{ document, created:
        false }` when the same source was already filed.


        **Access:** Service token only, from: `intake`. Browsers cannot call this.
      operationId: edrms_postDocuments
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - meta
                - file
              properties:
                meta:
                  type: string
                  description: "JSON: document type, fields, provenance… (must come before the file)"
                file:
                  type: string
                  format: binary
                  description: the scanned file
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - serviceToken: []
      x-access:
        services:
          - intake
    get:
      tags:
        - edrms · Documents
      summary: Search filed documents
      description: >-
        `{ items, total }`. `q` searches numbers, references and field values. Filter on any extracted field with `field.<key>=<value>`,
        e.g. `field.property=Erf 1873, Klein Windhoek`.


        **Access:** Signed-in user with any of `capture.view`, `verify.view`, `record.view`, `audit.view`, or a service token from `intake`,
        `land-records`, `audit`, `search`, `bpm`.
      operationId: edrms_getDocuments
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: docType
          in: query
          required: false
          schema:
            type: string
            enum:
              - deed_of_transfer
              - deed_of_sale
              - deed_of_grant
              - sg_diagram
              - mortgage_bond
              - other
        - name: batchId
          in: query
          required: false
          schema:
            type: string
            maxLength: 50
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
        - serviceToken: []
      x-access:
        anyOf:
          - capture.view
          - verify.view
          - record.view
          - audit.view
        services:
          - intake
          - land-records
          - audit
          - search
          - bpm
  /api/documents/{id}:
    get:
      tags:
        - edrms · Documents
      summary: One document with its fields, record metadata and version history
      description: >-
        **Access:** Signed-in user with any of `capture.view`, `verify.view`, `record.view`, `audit.view`, or a service token from `intake`,
        `land-records`, `audit`, `search`, `bpm`.
      operationId: edrms_getDocumentsId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
        - serviceToken: []
      x-access:
        anyOf:
          - capture.view
          - verify.view
          - record.view
          - audit.view
        services:
          - intake
          - land-records
          - audit
          - search
          - bpm
  /api/documents/{id}/amendments:
    post:
      tags:
        - edrms · Filing
      summary: Apply an approved amendment (bpm only)
      description: >-
        Creates a new sealed version; the previous one is kept. Users request changes through bpm: POST
        /processes/document-amendment/instances. The approver must differ from the requester. Returns `{ document, version }`.


        **Access:** Service token only, from: `bpm`. Browsers cannot call this.
      operationId: edrms_postDocumentsIdAmendments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - reason
                - expectedVersion
                - actor
                - approvedBy
              properties:
                reason:
                  type: string
                expectedVersion:
                  type: integer
                  minimum: 1
                changes:
                  type: array
                  items:
                    type: object
                    required:
                      - k
                      - v
                    properties:
                      k:
                        type: string
                      v:
                        type: string
                recordMetadata:
                  type: object
                actor:
                  type: object
                  required:
                    - id
                    - name
                  properties:
                    id:
                      type: string
                      minLength: 1
                    name:
                      type: string
                approvedBy:
                  type: object
                  required:
                    - id
                    - name
                  properties:
                    id:
                      type: string
                      minLength: 1
                    name:
                      type: string
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - serviceToken: []
      x-access:
        services:
          - bpm
  /api/documents/{id}/content:
    get:
      tags:
        - edrms · Documents
      summary: Short-lived link to the file
      description: >-
        `{ url, version, mimeType, sha256, expiresIn }`. `?version=` defaults to the current version. Open `url` directly (it needs no
        token): a presigned S3 URL, or GET /content/{id}/{n} for local storage.


        **Access:** Signed-in user with any of `capture.view`, `verify.view`, `record.view`, `audit.view`, or a service token from `intake`,
        `land-records`, `audit`, `search`, `bpm`.
      operationId: edrms_getDocumentsIdContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
        - serviceToken: []
      x-access:
        anyOf:
          - capture.view
          - verify.view
          - record.view
          - audit.view
        services:
          - intake
          - land-records
          - audit
          - search
          - bpm
  /api/documents/{id}/versions/{n}:
    get:
      tags:
        - edrms · Documents
      summary: One version of a document
      description: >-
        **Access:** Signed-in user with any of `capture.view`, `verify.view`, `record.view`, `audit.view`, or a service token from `intake`,
        `land-records`, `audit`, `search`, `bpm`.
      operationId: edrms_getDocumentsIdVersionsN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: "n"
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
        - serviceToken: []
      x-access:
        anyOf:
          - capture.view
          - verify.view
          - record.view
          - audit.view
        services:
          - intake
          - land-records
          - audit
          - search
          - bpm
  /api/documents/{id}/versions/{n}/verify:
    get:
      tags:
        - edrms · Documents
      summary: Integrity check of a version
      description: |-
        Recomputes the SHA-256 of the stored file and checks the seal: `{ intact, contentIntact, sealIntact, version, checkedAt, … }`.

        **Access:** Signed-in user with permission `audit.view`.
      operationId: edrms_getDocumentsIdVersionsNVerify
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: "n"
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - audit.view
  /api/documents/lookup:
    get:
      tags:
        - edrms · Documents
      summary: Find one document by a reference
      description: >-
        By `edrmsNo`, `instrumentRef` (e.g. T 2210/2008) or `sourceId` (the intake document id). 404 when there is none.


        **Access:** Signed-in user with any of `capture.view`, `verify.view`, `record.view`, `audit.view`, or a service token from `intake`,
        `land-records`, `audit`, `search`, `bpm`.
      operationId: edrms_getDocumentsLookup
      parameters:
        - name: edrmsNo
          in: query
          required: false
          schema:
            type: string
        - name: instrumentRef
          in: query
          required: false
          schema:
            type: string
        - name: sourceId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
        - serviceToken: []
      x-access:
        anyOf:
          - capture.view
          - verify.view
          - record.view
          - audit.view
        services:
          - intake
          - land-records
          - audit
          - search
          - bpm
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/auth/login or /api/auth/mfa (15 minutes; renew with POST /api/auth/refresh).
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Service-to-service token (typ "service"), signed with the shared JWT_SECRET. Not available to browsers.
    refreshCookie:
      type: apiKey
      in: cookie
      name: lrfe_rt
      description: httpOnly refresh cookie, path /api/auth, set at sign-in.
  schemas:
    Error:
      type: object
      required:
        - status
        - message
        - timestamp
        - path
      properties:
        status:
          type: string
          example: Forbidden
          description: HTTP status text
        message:
          type: string
          example: Not permitted
        details:
          description: Extra data on some errors, e.g. `{ perm }` on 403, `{ blockers }` when a document cannot be filed yet
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /users